+++ src/main.js
@@ -19,6 +19,7 @@ import { renderGuide } from './views/guide.js';
 import { renderStocksInPlay } from './views/stocksInPlay.js';
 import { renderAutomation } from './views/automation.js';
 import { renderAutomationGuide } from './views/automationGuide.js';
+import { renderAutoOrders } from './views/autoOrders.js';
 import { onForegroundMessage, isFCMSupported } from './data/messaging.js';
 
 // Theme + font-size attributes are pre-set by an inline script in index.html
@@ -71,6 +72,7 @@ function mountAppShell() {
   route('stocksinplay', renderStocksInPlay);
   route('automation',       renderAutomation);
   route('automation-guide', renderAutomationGuide);
+  route('auto-orders',      renderAutoOrders);
   defaultRoute('dashboard');
   const dispatch = start(main);